Allow all blueprints to be made under a route
