+++ blueprints/layout/index.js
@@ -1,6 +1,15 @@
+const InRoute = require('../in-route');
+
 module.exports = {
     description() {
 
         return 'generates a functional layout component';
+    },
+
+    fileMapTokens() {
+
+        return {
+            __root__: InRoute
+        };
     }
 };
