+++ blueprints/smart/index.js
@@ -1,3 +1,5 @@
+const InRoute = require('../in-route');
+
 module.exports = {
     description() {
 
@@ -7,6 +9,7 @@ module.exports = {
     fileMapTokens() {
 
         return {
+            __root__: InRoute,
             __smart__: (options) => {
 
                 return options.settings.getSetting('smartPath');
